add types as constants
