+++ splight_models/component.py
@@ -1,4 +1,10 @@
-from splight_models.constants import ComponentSize, RestartPolicy, LogginLevel, ComponentStatus
+from splight_models.constants import (
+    ComponentSize,
+    RestartPolicy,
+    LogginLevel,
+    ComponentStatus,
+    ComponentType
+)
 from splight_models.asset import Asset
 from splight_models.attribute import Attribute
 from splight_models.base import SplightBaseModel
@@ -95,11 +101,6 @@ class ComponentCommandStatus(str, Enum):
     SUCCESS = "succeeded"
     ERROR = "error"
 
-class ComponentType(str,Enum):
-    ALGORITHM = "algorithm"
-    NETWORK = "network"
-    CONNECTOR = "connector"
-    SIMULATOR = "simulator"
 
 class ComponentCommand(SplightBaseModel):
     id: Optional[str]
@@ -131,7 +132,7 @@ class BaseComponent(SplightBaseModel):
     name: Optional[str] = None
     version: str
     custom_types: Optional[List[CustomType]] = []
-    component_type: Optional[ComponentType] = None
+    component_type: Optional[ComponentType]
     input: Optional[List[InputParameter]] = []
     output: Optional[List[Output]] = []
     commands: Optional[List[Command]] = []
